Conditionally deleting old files in build scripts

+++ build_dict.sh
@@ -4,5 +4,5 @@ KANJIDIC=util/kanjidic
 EDICT=util/edict
 DICT=yomi_base/japanese/dictionary.db
 
-rm $DICT
+[ -f $DICT ] && rm $DICT
 util/compile.py --kanjidic $KANJIDIC --edict $EDICT $DICT
